requirements.txt
Pillow
discord